Initial commit after committing private info in the original git repo
This is the initial public version of this repo.  There is a missing
file here, local.py, which should contain the list of project_numbers
which you should check against to validate that this is one of your
instances.
I need to write a README for this repo to explain it.